code/requirements.txt
torch>=2.0
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2
pyyaml>=6.0
matplotlib>=3.6
tqdm>=4.65
